# openapi.yaml
openapi: 3.0.0
info:
  title: ShiftManager API
  version: 1.0.0
  description: |
    Comprehensive REST API для управления сменами сотрудников с интеграцией Telegram.
    
    ## Features
    
    - 👥 **Employee Management** - CRUD операции для сотрудников
    - ⏰ **Shift Management** - Планирование и отслеживание смен
    - ☕ **Break Management** - Управление перерывами
    - 🚨 **Violation System** - Автоматическое обнаружение нарушений
    - ⭐ **Rating System** - Динамический расчет рейтингов
    - 📱 **Telegram Integration** - WebApp и Bot API
    - 📊 **Analytics** - Статистика и отчеты
    
    ## Authentication
    
    API использует Supabase Auth для аутентификации.
    Для большинства endpoints требуется JWT token в заголовке Authorization.
    
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    
    - Standard API: 100 requests/minute
    - Auth endpoints: 5 attempts/15 minutes
    - Bot API: 120 requests/minute
  contact:
    name: API Support
    email: support@shiftmanager.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://your-domain.vercel.app
    description: Production server (Vercel)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Supabase Auth
    telegramAuth:
      type: apiKey
      in: header
      name: x-telegram-init-data
      description: Telegram WebApp initData for authentication
    botSecret:
      type: apiKey
      in: header
      name: x-bot-secret
      description: Bot API secret key
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            name:
              type: string
            message:
              type: string
            statusCode:
              type: integer
            code:
              type: string
            details:
              type: object
          description: Error details
      example:
        error:
          name: ValidationError
          message: Validation failed
          statusCode: 400
          code: VALIDATION_FAILED
          details:
            errors:
              - field: email
                message: Invalid email format
...
    Shift:
      type: object
      properties:
        id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        planned_start_at:
          type: string
          format: date-time
        planned_end_at:
          type: string
          format: date-time
        actual_start_at:
          type: string
          format: date-time
          nullable: true
        actual_end_at:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [scheduled, active, paused, completed, cancelled]
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      required:
        - id
        - employee_id
        - planned_start_at
        - planned_end_at
        - status
...
  parameters:
...
    ShiftId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Shift UUID
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /api/shifts/{id}:
    get:
      tags: [Shifts]
      summary: Get shift by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ShiftId'
      responses:
        '200':
          description: Shift found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Shifts]
      summary: Update shift
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ShiftId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                planned_start_at:
                  type: string
                  format: date-time
                planned_end_at:
                  type: string
                  format: date-time
                status:
                  type: string
                  enum: [scheduled, active, paused, completed, cancelled]
      responses:
        '200':
          description: Shift updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Shift'